Payment Converted into Cents

+++ database/migrations/2025_09_25_085116_create_therapists_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->json('slots');
             $table->json('days');
-            $table->integer('charges');
+            $table->float('charges');
             $table->string('specialization', 100);
             $table->timestamps();
         });
